Refactors inclusion
